# todos/docs/swagger.yaml
basePath: /
definitions:
  requests.NewTodo:
    properties:
      deadline:
        format: int64
        type: integer
      description:
        type: string
      title:
        type: string
    required:
    - title
    type: object
  responses.Todo:
    properties:
      deadline:
        type: integer
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      user_id:
        type: integer
    type: object
  responses.httpError:
    properties:
      error:
        type: object
    type: object
info:
  contact: {}
  title: Todos
  version: "1.0"
paths:
  /todos:
    post:
      parameters:
      - description: JWT-токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: Новая задача
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/requests.NewTodo'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.Todo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.httpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.httpError'
      summary: Создать новую задачу
      tags:
      - todos
  /todos/:id:
    put:
      parameters:
      - description: JWT-токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      - description: Новые значения задачи
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/requests.NewTodo'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Todo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.httpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.httpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.httpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.httpError'
      summary: Обновить существующую задачу
      tags:
      - todos
securityDefinitions:
  JWT-Token:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"